add integration tests for sync writer
