+++ ddtrace/internal/writer.py
@@ -193,7 +193,7 @@ class AgentWriter(_worker.PeriodicWorkerThread, TraceWriter):
         timeout=agent.DEFAULT_TIMEOUT,
         dogstatsd=None,
         report_metrics=False,
-        sync_flush_mode=False,
+        sync_mode=False,
     ):
         super(AgentWriter, self).__init__(interval=processing_interval, name=self.__class__.__name__)
         self.agent_url = agent_url
@@ -231,7 +231,7 @@ class AgentWriter(_worker.PeriodicWorkerThread, TraceWriter):
         self._report_metrics = report_metrics
         self._metrics_reset()
         self._drop_sma = SimpleMovingAverage(DEFAULT_SMA_WINDOW)
-        self._sync_flush_mode = sync_flush_mode
+        self._sync_mode = sync_mode
 
     def _metrics_dist(self, name, count=1, tags=None):
         self._metrics[name]["count"] += count
@@ -265,7 +265,7 @@ class AgentWriter(_worker.PeriodicWorkerThread, TraceWriter):
         writer = self.__class__(
             agent_url=self.agent_url,
             priority_sampler=self._priority_sampler,
-            sync_flush_mode=self._sync_flush_mode,
+            sync_mode=self._sync_mode,
         )
         writer._encoder = self._encoder
         writer._headers = self._headers
@@ -349,7 +349,7 @@ class AgentWriter(_worker.PeriodicWorkerThread, TraceWriter):
         # Start the AgentWriter on first write.
         # Starting it earlier might be an issue with gevent, see:
         # https://github.com/DataDog/dd-trace-py/issues/1192
-        if self.started is False and self._sync_flush_mode is False:
+        if self.started is False and self._sync_mode is False:
             with self._started_lock:
                 if self.started is False:
                     self.start()
@@ -386,7 +386,7 @@ class AgentWriter(_worker.PeriodicWorkerThread, TraceWriter):
             else:
                 self._metrics_dist("buffer.accepted.traces", 1)
                 self._metrics_dist("buffer.accepted.spans", len(spans))
-                if self._sync_flush_mode:
+                if self._sync_mode:
                     self.flush_queue()
 
     def flush_queue(self, raise_exc=False):
